[Api] Moving Api classes to a new Manager namespace for better organization.
